scripts/requirements.txt
python-docx>=1.1.0
weasyprint>=62.0
markdown>=3.7
Pygments>=2.18
